refactor(*): update package structure
