+++ lib/src/data_state_enum.dart
